requirements.txt
Django>=1.8
djangorestframework>=3.1.1
MySQL-python>=1.2.5
Fabric>=1.10.2
django-extensions>=1.5.5
pygraphviz